+++ sprockets/mixins/correlation/__init__.py
@@ -1,9 +1,12 @@
 try:
-    from .mixins import HandlerMixin
+    from .mixins import CorrelationAdapter, HandlerMixin
 
 # the following makes importing the module possible when tornado
 # is not installed in the active environment
 except ImportError as exc:  # pragma no cover
+    def CorrelationAdapter(*args, **kwargs):
+        raise exc
+
     def HandlerMixin(*args, **kwargs):
         raise exc